[alena3][webcall-app] added presence into chat
git-svn-id: https://svn.rcslabs.ru/subversion/rcslabs/SOURCES/trunk/alena3/webcall-app@6487 44e08e39-4b91-0410-a4d4-833ecb1b66d7

+++ src/main/java/com/rcslabs/chat/ChatAuthController.java
@@ -26,6 +26,9 @@ public class ChatAuthController extends AbstractAuthController {
 
     @Override
     public void closeSession(String sessionId) {
+        ISession session = storage.get(sessionId);
+        if(null == session) return;
         storage.delete(sessionId);
+        super.onSessionClosed(session);
     }
 }
